A bit more clean up
