+++ extension/ts/shared-resources/types/bergamot.types.d.ts
@@ -1,11 +1,3 @@
-export interface DetectedLanguageResults {
-  confident: boolean;
-  language: string;
-  languages: {
-    languageCode: string;
-    percent: number;
-  }[];
-}
 export interface FrameInfo {
   windowId: number;
   tabId: number;
